prac 7 - real final

+++ Prac 7/main.py
@@ -4,6 +4,10 @@ from flask_cors import CORS
 import json
 import models
 
+'''
+NOTE: for this prac, just open the html page
+'''
+
 app = Flask(__name__)
 CORS(app, support_credentials=True)
 
@@ -12,49 +16,44 @@ app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = True
 db = SQLAlchemy(app)
 
 
-@app.route('/', methods=['GET'])
-def query_records():
-    ''' SELECT '''
-    results = models.get_products()
-    products = []
-    for r in results:
-        products.append({
-            "id": r.id,
-            "name": r.name,
-            "price": r.price
-        })
-    return {"products": products}
+@app.route('/', methods=['GET', "POST", "PUT", "DELETE"])
+def product_page():
+    if request.method == "GET":
+        ''' SELECT '''
+        results = models.get_products()
+        products = []
+        for r in results:
+            products.append({
+                "id": r.id,
+                "name": r.name,
+                "price": r.price
+            })
+        return {"products": products}
 
+    elif request.method == "POST":
+        ''' CREATE '''
+        data = request.get_json()
+        name = str(data["name"])
+        price = float(data["price"])
+        res = models.create_product(name, price)
+        return str(res)
 
-@app.route('/', methods=['POST'])
-def create_record():
-    ''' CREATE '''
-    data = request.get_json()
-    name = str(data["name"])
-    price = float(data["price"])
-    res = models.create_product(name, price)
-    return str(res)
+    elif request.method == "PUT":
+        ''' UPDATE '''
+        data = request.get_json()
+        _id = int(data["id"])
+        name = data["name"]
+        price = float(data["price"])
+        res = models.update_product(_id, name, price)
+        return jsonify(res)
 
-
-@app.route('/', methods=['PUT'])
-def update_record():
-    ''' UPDATE '''
-    data = request.get_json()
-    _id = int(data["id"])
-    name = data["name"]
-    price = float(data["price"])
-    res = models.update_product(_id, name, price)
-    return jsonify(res)
-
-
-@app.route('/', methods=['DELETE'])
-def delete_record():
-    ''' DELETE '''
-    data = request.get_json()
-    _id = int(data["id"])
-    #print(data)
-    res = models.delete_product(_id)
-    return jsonify(res)
+    elif request.method == "DELETE":
+        ''' DELETE '''
+        data = request.get_json()
+        _id = int(data["id"])
+        #print(data)
+        res = models.delete_product(_id)
+        return jsonify(res)
 
 
 if __name__ == "__main__":
